requirements.txt
adrf==0.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.0.0
appnope==0.1.4
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
babel==2.16.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blessed==1.20.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.5
cloudinary==1.33.0
comm==0.2.2
contourpy==1.3.0
coreschema==0.0.4
cryptography==41.0.4
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
Django==5.1.4
django-after-response==0.2.2
django-autoslug==1.9.9
django-cities-light==3.9.2
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
dnspython==2.7.0
docutils==0.21.2
drf-spectacular==0.26.3
dxh_py==0.0.5
editor==1.6.6
executing==2.1.0
fastapi==0.115.4
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
html2image==2.0.5
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.4
imagesize==1.4.1
importlib_metadata==8.5.0
inflection==0.5.1
iniconfig==2.0.0
inquirer==3.4.0
instigator_py==0.0.8
ipykernel==6.29.5
ipython==8.29.0
itypes==1.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.7.0
joblib==1.4.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.5.0
kiwisolver==1.4.7
litellm==1.51.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
MouseInfo==0.1.3
msgpack==1.0.5
multidict==6.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
nh3==0.2.18
nltk==3.9.1
numpy==2.1.2
open-interpreter==0.4.3
openai==0.27.8
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.2.0
progressbar2==4.2.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==5.9.8
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.7.0
pymongo==4.10.1
PyMsgBox==1.0.9
PyMuPDF==1.24.13
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.2.0
pyperclip==1.9.0
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
python-utils==3.7.0
pytube==15.0.0
pytweening==1.2.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==26.2.0
readchar==4.2.0
readme_renderer==44.0
redis==4.6.0
referencing==0.29.1
regex==2024.9.11
renamering==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.3
rpds-py==0.8.10
rsa==4.9
rubicon-objc==0.4.9
runs==1.2.2
selenium==4.26.1
Send2Trash==1.8.3
setuptools==75.5.0
setuptools-scm==8.1.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.11.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
termcolor==2.3.0
text-unidecode==1.3
tiktoken==0.7.0
tokenizers==0.20.1
tokentrim==0.1.13
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
twine==5.1.1
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.0
watchfiles==0.20.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
wget==3.2
wheel==0.45.0
whitenoise==6.5.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.17.1
yaspin==3.1.0
yfinance==0.2.50
yt-dlp==2024.11.18
zipp==3.20.2
